Delete README.md
I have an existing worship music program which permits volume edits of 10 DB each.  This is too wide of a swing so being able to control the volume in 1 Db increments would be very beneficial.  Also; the addition of a digital volume meter would permit the cue editor to see the before and after effect of the volume edit.